Start extracting Logical Excel functions into separate groups in dedicated classes
