+++ src/PhpSpreadsheet/Calculation/Logical.php
@@ -2,6 +2,8 @@
 
 namespace PhpOffice\PhpSpreadsheet\Calculation;
 
+use PhpOffice\PhpSpreadsheet\Calculation\Logical\Boolean;
+
 class Logical
 {
     /**
@@ -12,11 +14,13 @@ class Logical
      * Excel Function:
      *        =TRUE()
      *
+     * @Deprecated Use the TRUE() method in the Logical\Boolean class instead
+     *
      * @return bool True
      */
-    public static function true()
+    public static function true(): bool
     {
-        return true;
+        return Boolean::true();
     }
 
     /**
@@ -27,37 +31,13 @@ class Logical
      * Excel Function:
      *        =FALSE()
      *
+     * @Deprecated Use the FALSE() method in the Logical\Boolean class instead
+     *
      * @return bool False
      */
-    public static function false()
+    public static function false(): bool
     {
-        return false;
-    }
-
-    private static function countTrueValues(array $args)
-    {
-        $returnValue = 0;
-
-        foreach ($args as $arg) {
-            // Is it a boolean value?
-            if (is_bool($arg)) {
-                $returnValue += $arg;
-            } elseif ((is_numeric($arg)) && (!is_string($arg))) {
-                $returnValue += ((int) $arg != 0);
-            } elseif (is_string($arg)) {
-                $arg = strtoupper($arg);
-                if (($arg == 'TRUE') || ($arg == Calculation::getTRUE())) {
-                    $arg = true;
-                } elseif (($arg == 'FALSE') || ($arg == Calculation::getFALSE())) {
-                    $arg = false;
-                } else {
-                    return Functions::VALUE();
-                }
-                $returnValue += ($arg != 0);
-            }
-        }
-
-        return $returnValue;
+        return Boolean::false();
     }
 
     /**
@@ -73,8 +53,10 @@ class Logical
      *
      *        Boolean arguments are treated as True or False as appropriate
      *        Integer or floating point arguments are treated as True, except for 0 or 0.0 which are False
-     *        If any argument value is a string, or a Null, the function returns a #VALUE! error, unless the string holds
-     *            the value TRUE or FALSE, in which case it is evaluated as the corresponding boolean value
+     *        If any argument value is a string, or a Null, the function returns a #VALUE! error, unless the string
+     *            holds the value TRUE or FALSE, in which case it is evaluated as the corresponding boolean value
+     *
+     * @Deprecated Use the logicalAnd() method in the Logical\Operations class instead
      *
      * @param mixed ...$args Data values
      *
@@ -82,23 +64,7 @@ class Logical
      */
     public static function logicalAnd(...$args)
     {
-        $args = Functions::flattenArray($args);
-
-        if (count($args) == 0) {
-            return Functions::VALUE();
-        }
-
-        $args = array_filter($args, function ($value) {
-            return $value !== null || (is_string($value) && trim($value) == '');
-        });
-        $argCount = count($args);
-
-        $returnValue = self::countTrueValues($args);
-        if (is_string($returnValue)) {
-            return $returnValue;
-        }
-
-        return ($returnValue > 0) && ($returnValue == $argCount);
+        return Logical\Operations::logicalAnd(...$args);
     }
 
     /**
@@ -114,8 +80,10 @@ class Logical
      *
      *        Boolean arguments are treated as True or False as appropriate
      *        Integer or floating point arguments are treated as True, except for 0 or 0.0 which are False
-     *        If any argument value is a string, or a Null, the function returns a #VALUE! error, unless the string holds
-     *            the value TRUE or FALSE, in which case it is evaluated as the corresponding boolean value
+     *        If any argument value is a string, or a Null, the function returns a #VALUE! error, unless the string
+     *            holds the value TRUE or FALSE, in which case it is evaluated as the corresponding boolean value
+     *
+     * @Deprecated Use the logicalOr() method in the Logical\Operations class instead
      *
      * @param mixed $args Data values
      *
@@ -123,29 +91,15 @@ class Logical
      */
     public static function logicalOr(...$args)
     {
-        $args = Functions::flattenArray($args);
-
-        if (count($args) == 0) {
-            return Functions::VALUE();
-        }
-
-        $args = array_filter($args, function ($value) {
-            return $value !== null || (is_string($value) && trim($value) == '');
-        });
-
-        $returnValue = self::countTrueValues($args);
-        if (is_string($returnValue)) {
-            return $returnValue;
-        }
-
-        return $returnValue > 0;
+        return Logical\Operations::logicalOr(...$args);
     }
 
     /**
      * LOGICAL_XOR.
      *
      * Returns the Exclusive Or logical operation for one or more supplied conditions.
-     * i.e. the Xor function returns TRUE if an odd number of the supplied conditions evaluate to TRUE, and FALSE otherwise.
+     * i.e. the Xor function returns TRUE if an odd number of the supplied conditions evaluate to TRUE,
+     *      and FALSE otherwise.
      *
      * Excel Function:
      *        =XOR(logical1[,logical2[, ...]])
@@ -155,8 +109,10 @@ class Logical
      *
      *        Boolean arguments are treated as True or False as appropriate
      *        Integer or floating point arguments are treated as True, except for 0 or 0.0 which are False
-     *        If any argument value is a string, or a Null, the function returns a #VALUE! error, unless the string holds
-     *            the value TRUE or FALSE, in which case it is evaluated as the corresponding boolean value
+     *        If any argument value is a string, or a Null, the function returns a #VALUE! error, unless the string
+     *            holds the value TRUE or FALSE, in which case it is evaluated as the corresponding boolean value
+     *
+     * @Deprecated Use the logicalXor() method in the Logical\Operations class instead
      *
      * @param mixed $args Data values
      *
@@ -164,22 +120,7 @@ class Logical
      */
     public static function logicalXor(...$args)
     {
-        $args = Functions::flattenArray($args);
-
-        if (count($args) == 0) {
-            return Functions::VALUE();
-        }
-
-        $args = array_filter($args, function ($value) {
-            return $value !== null || (is_string($value) && trim($value) == '');
-        });
-
-        $returnValue = self::countTrueValues($args);
-        if (is_string($returnValue)) {
-            return $returnValue;
-        }
-
-        return $returnValue % 2 == 1;
+        return Logical\Operations::logicalXor(...$args);
     }
 
     /**
@@ -194,8 +135,10 @@ class Logical
      *
      *        Boolean arguments are treated as True or False as appropriate
      *        Integer or floating point arguments are treated as True, except for 0 or 0.0 which are False
-     *        If any argument value is a string, or a Null, the function returns a #VALUE! error, unless the string holds
-     *            the value TRUE or FALSE, in which case it is evaluated as the corresponding boolean value
+     *        If any argument value is a string, or a Null, the function returns a #VALUE! error, unless the string
+     *            holds the value TRUE or FALSE, in which case it is evaluated as the corresponding boolean value
+     *
+     * @Deprecated Use the NOT() method in the Logical\Operations class instead
      *
      * @param mixed $logical A value or expression that can be evaluated to TRUE or FALSE
      *
@@ -203,20 +146,7 @@ class Logical
      */
     public static function NOT($logical = false)
     {
-        $logical = Functions::flattenSingleValue($logical);
-
-        if (is_string($logical)) {
-            $logical = strtoupper($logical);
-            if (($logical == 'TRUE') || ($logical == Calculation::getTRUE())) {
-                return false;
-            } elseif (($logical == 'FALSE') || ($logical == Calculation::getFALSE())) {
-                return true;
-            }
-
-            return Functions::VALUE();
-        }
-
-        return !$logical;
+        return Logical\Operations::NOT($logical);
     }
 
     /**
